chatbots/pyproject.toml
[project]
name = "chatbothub"
version = "1.0.0"
description = "Agents IA - chatbots pour MS Teams"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "botbuilder-integration-aiohttp>=4.17.0",
    "botbuilder-core>=4.17.0",
    "botbuilder-dialogs>=4.17.0",
    "aiohttp>=3.12.15",
    "openai>=1.107.1",
    "pyfiglet>=1.0.4",
    "google-genai>=1.36.0",
    "ddtrace>=3.13.2",
    "pandas>=2.3.2",
    "matplotlib>=3.10.6",
    "pytest-asyncio>=1.1.0",
    "azure-communication-callautomation>=1.5.0",
    "azure-identity>=1.25.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=8.4.2",
    "pytest-asyncio",
]
